api/requirements.txt
requests
flask
flask_cors
pillow